Fix error response mapper not propagating http error responses not translated to an exception
